+++ client/src/agent.js
@@ -1,10 +1,26 @@
+const jwtDecode = require("jwt-decode");
+const qs = require("qs");
 import _superagent from "superagent";
 import superagentPromise from "superagent-promise";
-import CommonStore from "./stores/CommonStore";
 
 const superagent = superagentPromise(_superagent, global.Promise);
 
 const API_ROOT = "/api";
+const QUERY_PARAMETERS = window.location.search;
+
+/**
+ * @param {string} item
+ * @return {(string|boolean)}
+ */
+function parseQueryParams(item) {
+  if (!item) return false;
+  try {
+    return qs.parse(QUERY_PARAMETERS, { ignoreQueryPrefix: true })[item];
+  } catch (reason) {
+    console.error(reason);
+  }
+  return false;
+}
 
 const handleErrors = err => {
   return err;
@@ -13,8 +29,14 @@ const handleErrors = err => {
 const responseBody = res => res.body;
 
 const tokenPlugin = req => {
-  if (CommonStore.jwt) {
-    req.set("Authorization", `Bearer ${CommonStore.jwt}`);
+  let jwt;
+  const token = parseQueryParams("token");
+  try {
+    jwtDecode(token);
+    jwt = token;
+    req.set("Authorization", `Bearer ${jwt}`);
+  } catch (reason) {
+    console.error(reason);
   }
 };
 
@@ -28,19 +50,15 @@ const requests = {
   }
 };
 
-const Context = {
-  get: () => {
-    return requests.get(`/context/`);
-  }
-};
-
 const Canvas = {
   status: () => {
     return requests.get(`/canvas-status/`);
   }
 };
 
+const getContext = () => requests.get(`/context/`);
+
 export default {
   Canvas,
-  Context
+  getContext
 };